+++ Xpand/Xpand.ExpressApp.Modules/AdditionalViewControlsProvider/Security/AdditionalViewControlsPermission.cs
@@ -26,7 +26,7 @@ namespace Xpand.ExpressApp.AdditionalViewControlsProvider.Security {
         public Color? ForeColor { get; set; }
         public FontStyle? FontStyle { get; set; }
         public int? Height { get; set; }
-        public float? FontSize { get; set; }
+        public int? FontSize { get; set; }
         #endregion
         public override IPermission Copy() {
             return new AdditionalViewControlsPermission();
